wazo-load-pilot: move python files into its own package
