+++ tox.ini
@@ -24,7 +24,7 @@ deps =
 [testenv:pilot]
 change_dir = wazo_load_pilot
 commands =
-    pytest -v --pyargs plugins/
+    pytest -v --pyargs wlpd/plugins
 
 deps =
     -rwazo_load_pilot/requirements.txt
