--- wazo_load_pilot/plugins/__init__.py
@@ -1,4 +1,0 @@
-# Copyright 2023 The Wazo Authors  (see the AUTHORS file)
-# SPDX-License-Identifier: GPL-3.0-or-later
-
-__all__ = ['plugins']
