+++ wazo_load_pilot/wlpd/plugins/fleet/http.py
