+++ wazo_load_pilot/wlpd/plugins/fleet/plugin.py
