+++ wazo_load_pilot/wlpd/plugins/pilot/commands.py
