+++ wazo_load_pilot/wlpd/tests/certificate.csr
