+++ wazo_load_pilot/wlpd/tests/private.key
